api/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kvasir-api"
version = "0.1.0"
description = "API for Kvasir"
authors = [
    { name = "Eirik Drage Steen", email = "eirik.drage.steen@gmail.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy",
    "argparse",
    "matplotlib",
    "seaborn",
    "pandas",
    "scipy",
    "openai",
    "pydantic",
    "instructor",
    "asyncpg",
    "sqlalchemy[asyncio]",
    "fastapi[standard]",
    "uvicorn",
    "pydantic-ai",
    "psycopg2-binary",
    "sqlalchemy-utils",
    "PyJWT",
    "passlib",
    "bcrypt==4.3.0",
    "httpx",
    "aiofiles",
    "alembic",
    "redis",
    "statsmodels",
    "markdown2",
    "boto3",
    "openpyxl",
    "docker",
    "pylint",
    "taskiq",
    "taskiq-redis",
    "pgvector",
    "google-generativeai",
    "cryptography",
    "jsonschema",
    "weasyprint",
    "pyecharts",
    "snapshot-selenium",
    "pyarrow"
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["kvasir_api"]